Keep trailing punctuation attached to its value when wrapping
Each field and its trailing punctuation (a separator, or the closing
affix for the last field) were separate FlowLayout children, so a
comma or the closing paren could wrap onto the next line by itself,
orphaned ahead of the value it belongs to. Group each field with its
trailing punctuation into one atomic HStack wrap unit instead.

Also extract ScrollValueAdapter/ScrubbableColorField/ScrubTextField
into their own file — EditableColorValue.swift was right at
SwiftLint's 1000-line hard limit and this change would have tipped
it over.
