+++ Pika/Views/EditableColorValue.swift
@@ -137,23 +137,30 @@ struct EditableColorValue: View {
         FlowLayout(maxLines: Int(maxLines)) {
             affix(layout.leading, size: size)
             ForEach(Array(layout.components.enumerated()), id: \.offset) { index, component in
-                ColorComponentField(
-                    text: binding(for: index, layout: layout),
-                    component: component,
-                    index: index,
-                    uiColor: uiColor,
-                    fontSize: size,
-                    focusedIndex: $focusedIndex,
-                    onSubmit: commitEditing,
-                    onCancel: revertEditing,
-                    onDragBegin: { beginDragSession(index: index, layout: layout) },
-                    onDragEnd: { finishDragOrScrollSession(index: index) }
-                )
-                if index < layout.separators.count {
-                    affix(layout.separators[index], size: size)
+                // Grouped with its trailing punctuation (the separator after it, or the closing
+                // affix for the last one) into one atomic wrap unit — otherwise `FlowLayout`
+                // could wrap that punctuation onto the next line by itself, orphaned ahead of
+                // the value it actually belongs to.
+                HStack(spacing: 0) {
+                    ColorComponentField(
+                        text: binding(for: index, layout: layout),
+                        component: component,
+                        index: index,
+                        uiColor: uiColor,
+                        fontSize: size,
+                        focusedIndex: $focusedIndex,
+                        onSubmit: commitEditing,
+                        onCancel: revertEditing,
+                        onDragBegin: { beginDragSession(index: index, layout: layout) },
+                        onDragEnd: { finishDragOrScrollSession(index: index) }
+                    )
+                    if index < layout.separators.count {
+                        affix(layout.separators[index], size: size)
+                    } else {
+                        affix(layout.trailing, size: size)
+                    }
                 }
             }
-            affix(layout.trailing, size: size)
         }
         // An explicit width, not `.frame(maxWidth: .infinity)`: a plain flexible frame was found
         // to sometimes only ever be queried for its *ideal* size in this view's position in the
@@ -400,7 +407,7 @@ struct EditableColorValue: View {
 /// click-to-focus — a raw mouseDown on a real `TextField` always focuses/selects it immediately,
 /// before any gesture recognizer gets a chance to see the drag. Owning `mouseDown` directly is
 /// the only way to decide click-vs-drag *before* anything focuses.
-private struct ColorComponentField: View {
+struct ColorComponentField: View {
     @Binding var text: String
     let component: ColorComponent
     let index: Int
@@ -550,7 +557,7 @@ private struct ColorComponentField: View {
         onDragEnd()
     }
 
-    fileprivate static func formattedDragValue(_ value: Double, kind: ComponentKind) -> String {
+    static func formattedDragValue(_ value: Double, kind: ComponentKind) -> String {
         switch kind {
         case .hex:
             return ""
@@ -559,442 +566,5 @@ private struct ColorComponentField: View {
         case .decimal:
             return CGFloat(value).strippedDecimalString(maxDecimalPlaces: 4)
         }
-    }
-}
-
-/// Captures two-finger trackpad scroll events landing within the wrapped view's bounds and
-/// reports vertical delta/end, mirroring `HorizontalScrollWheelAdapter`'s local-monitor approach
-/// so a scrub field can be nudged the same way click-drag nudges it.
-private struct ScrollValueAdapter: NSViewRepresentable {
-    let isEnabled: Bool
-    let onScroll: (CGFloat) -> Void
-    let onScrollEnd: () -> Void
-
-    func makeNSView(context _: Context) -> ScrollCaptureView {
-        let view = ScrollCaptureView()
-        view.isEnabled = isEnabled
-        view.onScroll = onScroll
-        view.onScrollEnd = onScrollEnd
-        return view
-    }
-
-    func updateNSView(_ view: ScrollCaptureView, context _: Context) {
-        view.isEnabled = isEnabled
-        view.onScroll = onScroll
-        view.onScrollEnd = onScrollEnd
-    }
-
-    final class ScrollCaptureView: NSView {
-        var isEnabled = true
-        var onScroll: ((CGFloat) -> Void)?
-        var onScrollEnd: (() -> Void)?
-        private var monitor: Any?
-        private var isScrolling = false
-
-        override func viewDidMoveToWindow() {
-            super.viewDidMoveToWindow()
-            if window != nil, monitor == nil {
-                monitor = NSEvent.addLocalMonitorForEvents(matching: .scrollWheel) { [weak self] event in
-                    self?.handle(event) ?? event
-                }
-            } else if window == nil, let monitor {
-                NSEvent.removeMonitor(monitor)
-                self.monitor = nil
-            }
-        }
-
-        deinit {
-            if let monitor { NSEvent.removeMonitor(monitor) }
-        }
-
-        private func handle(_ event: NSEvent) -> NSEvent? {
-            guard isEnabled, event.window === window, let window else { return event }
-
-            // Once we own an in-progress scroll, keep tracking it no matter where the cursor
-            // goes: trackpad momentum keeps delivering events (often with a final `.ended`
-            // phase) after the user's fingers leave the trackpad, and by then the cursor has
-            // frequently drifted off this field. Gating termination on `bounds.contains` meant
-            // that final event was silently dropped, `onScrollEnd` never fired, and the parent's
-            // edit session leaked open — so the *next* field touched would find a stale
-            // "already editing" session and force-focus itself.
-            if isScrolling {
-                if event.phase == .ended || event.phase == .cancelled || event.momentumPhase == .ended {
-                    isScrolling = false
-                    onScrollEnd?()
-                    return event
-                }
-                guard event.hasPreciseScrollingDeltas else { return event }
-                let deltaY = event.scrollingDeltaY
-                guard deltaY != 0 else { return event }
-                onScroll?(deltaY)
-                return nil
-            }
-
-            let pointInSelf = convert(event.locationInWindow, from: nil)
-            guard bounds.contains(pointInSelf) else { return event }
-
-            // Only intervene for trackpad/Magic Mouse gesture scrolling, which carries
-            // phase/precise deltas; classic scroll wheels should keep their default behaviour.
-            guard event.hasPreciseScrollingDeltas else { return event }
-            guard event.phase != .ended, event.phase != .cancelled, event.momentumPhase != .ended else { return event }
-
-            let deltaY = event.scrollingDeltaY
-            guard deltaY != 0 else { return event }
-            isScrolling = true
-            onScroll?(deltaY)
-            return nil
-        }
-    }
-}
-
-/// Wraps `ScrubTextField` (below) for SwiftUI. Owns focus explicitly via `isFocused`/
-/// `onFocusChange` rather than `@FocusState`/`.focused()` (which don't bridge to a custom
-/// `NSViewRepresentable`), synced through the field's own `NSTextFieldDelegate` callbacks so it
-/// stays correct however focus changes — click, Tab, or a programmatic request.
-private struct ScrubbableColorField: NSViewRepresentable {
-    @Binding var text: String
-    let fontSize: CGFloat
-    let textColor: NSColor
-    let isDraggable: Bool
-    let range: ClosedRange<Double>?
-    let kind: ComponentKind
-    let isFocused: Bool
-    let onFocusChange: (Bool) -> Void
-    let onSubmit: () -> Void
-    let onCancel: () -> Void
-    /// Fired when a click-drag-to-scrub gesture starts/ends, so the parent can wrap it in the
-    /// same live-preview/commit session used for typed edits.
-    let onDragBegin: () -> Void
-    let onDragEnd: () -> Void
-    /// Fired with +1/-1 for Up/Down arrow keys, `nil` for non-draggable (hex) fields.
-    let onStep: ((CGFloat) -> Void)?
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(text: $text, onSubmit: onSubmit, onCancel: onCancel, onStep: onStep)
-    }
-
-    func makeNSView(context: Context) -> ScrubTextField {
-        let field = ScrubTextField()
-        field.isBordered = false
-        field.drawsBackground = false
-        field.focusRingType = .none
-        // Left, not center: the field's frame is already sized to fit its text almost exactly
-        // (`intrinsicContentSize`, below), so centring only has ~2pt of slack to distribute. The
-        // static cell measures that slack with `NSString.size(withAttributes:)`; the live field
-        // editor lays the same string out via TextKit, which can measure it a device pixel
-        // narrower/wider — enough to visibly shift the centred text the moment editing begins.
-        // Left alignment anchors the text to the same edge under both renderers, so it doesn't move.
-        field.alignment = .left
-        field.usesSingleLineMode = true
-        field.cell?.wraps = false
-        field.delegate = context.coordinator
-        field.font = NSFont.systemFont(ofSize: fontSize, weight: .regular)
-        field.stringValue = text
-        return field
-    }
-
-    func updateNSView(_ nsView: ScrubTextField, context: Context) {
-        context.coordinator.text = $text
-        // Driven off `becomeFirstResponder`/`resignFirstResponder` directly rather than the
-        // `NSTextFieldDelegate` controlTextDidBeginEditing/EndEditing notifications: a click that
-        // lands on a not-yet-focused field goes through AppKit's own private pre-focus path
-        // (`NSWindow._handleMouseDownEvent:` → `NSTextFieldCell _selectOrEdit:…`) *before* this
-        // view's `mouseDown` override ever runs, and that path never posts the notifications the
-        // delegate relies on — so `focusedIndex` silently never got set, and the focus outline
-        // never appeared. The responder overrides fire reliably however focus changes.
-        nsView.onFocusChange = onFocusChange
-        context.coordinator.onStep = onStep
-        if nsView.font?.pointSize != fontSize {
-            nsView.font = NSFont.systemFont(ofSize: fontSize, weight: .regular)
-            nsView.invalidateIntrinsicContentSize()
-        }
-        nsView.textColor = textColor
-        nsView.isDraggable = isDraggable
-        nsView.range = range
-        nsView.kind = kind
-        nsView.onDragBegin = onDragBegin
-        nsView.onDragChanged = { [weak nsView] newValue in
-            guard let nsView else { return }
-            nsView.stringValue = ColorComponentField.formattedDragValue(newValue, kind: kind)
-            // Setting `stringValue` on a field that still has a live field editor attached
-            // (see `beginDrag`'s comment) resets the editor's selection to select-all every
-            // time — so every step of the drag re-selects the newly-set text, and whichever
-            // step happens to be the last one is what's left visibly selected once the drag
-            // ends. Collapse it again after every step, not just the first.
-            if let editor = nsView.currentEditor() {
-                let length = (editor.string as NSString).length
-                editor.selectedRange = NSRange(location: length, length: 0)
-            }
-            text = nsView.stringValue
-        }
-        nsView.onDragEnd = onDragEnd
-
-        if nsView.stringValue != text {
-            nsView.stringValue = text
-            nsView.invalidateIntrinsicContentSize()
-            // A programmatic change (drag/scroll, or an external colour landing mid-edit) while
-            // this field is still first responder — keep the caret collapsed at the end rather
-            // than whatever AppKit does by default when `stringValue` changes underneath it.
-            if let editor = nsView.currentEditor() as? NSTextView {
-                let length = (text as NSString).length
-                editor.selectedRange = NSRange(location: length, length: 0)
-            }
-        }
-
-        let editorIsActive = nsView.currentEditor() != nil && nsView.window?.firstResponder === nsView.currentEditor()
-        if isFocused, !editorIsActive {
-            nsView.window?.makeFirstResponder(nsView)
-        } else if !isFocused, editorIsActive {
-            nsView.window?.makeFirstResponder(nil)
-        }
-    }
-
-    final class Coordinator: NSObject, NSTextFieldDelegate {
-        var text: Binding<String>
-        let onSubmit: () -> Void
-        let onCancel: () -> Void
-        var onStep: ((CGFloat) -> Void)?
-
-        init(
-            text: Binding<String>, onSubmit: @escaping () -> Void, onCancel: @escaping () -> Void,
-            onStep: ((CGFloat) -> Void)?
-        ) {
-            self.text = text
-            self.onSubmit = onSubmit
-            self.onCancel = onCancel
-            self.onStep = onStep
-        }
-
-        func controlTextDidChange(_ obj: Notification) {
-            guard let field = obj.object as? ScrubTextField else { return }
-            // Clearing a ranged field (select-all + delete, or backspacing the last digit)
-            // leaves it both empty and invalid — effectively a dead end, since an empty
-            // `.fixedSize()` field also collapses to no width, hiding the invalid-state dashes
-            // that would otherwise show. Snap it to the component's lowest value instead, and
-            // select it, so the field stays visible, valid, and ready to be typed straight over
-            // — mirroring what a fresh click-to-select does. Hex has no natural minimum
-            // (`range` is nil), so it keeps the plain empty/invalid state.
-            if field.stringValue.isEmpty, let range = field.range {
-                let lowest = ColorComponentField.formattedDragValue(range.lowerBound, kind: field.kind)
-                field.stringValue = lowest
-                text.wrappedValue = lowest
-                if let editor = field.currentEditor() {
-                    editor.selectedRange = NSRange(location: 0, length: (lowest as NSString).length)
-                }
-                return
-            }
-            text.wrappedValue = field.stringValue
-        }
-
-        func control(_: NSControl, textView _: NSTextView, doCommandBy commandSelector: Selector) -> Bool {
-            if commandSelector == #selector(NSResponder.insertNewline(_:)) {
-                onSubmit()
-                return true
-            }
-            if commandSelector == #selector(NSResponder.cancelOperation(_:)) {
-                onCancel()
-                return true
-            }
-            if commandSelector == #selector(NSResponder.moveUp(_:)), let onStep {
-                onStep(1)
-                return true
-            }
-            if commandSelector == #selector(NSResponder.moveDown(_:)), let onStep {
-                onStep(-1)
-                return true
-            }
-            return false
-        }
-    }
-}
-
-/// A plain `NSTextField` subclass that owns its own `mouseDown`, so click-vs-drag is resolved
-/// *before* anything can focus — no second view or event monitor racing the field's native click
-/// handling. A resolved drag never touches first-responder status at all, so the field just
-/// displays a changing string like a label while scrubbing (no editor, no selection, no caret).
-/// A resolved click focuses normally; `becomeFirstResponder` then deterministically selects the
-/// whole value in the same call stack (matching the click-to-select-all behaviour of a spreadsheet
-/// cell), rather than reacting to AppKit's own click-positions-the-caret behaviour after the fact.
-private final class ScrubTextField: NSTextField {
-    var isDraggable = false {
-        didSet { window?.invalidateCursorRects(for: self) }
-    }
-
-    var range: ClosedRange<Double>?
-    var kind: ComponentKind = .integer
-    var onDragBegin: (() -> Void)?
-    var onDragChanged: ((Double) -> Void)?
-    var onDragEnd: (() -> Void)?
-    /// Reports true/false as this field becomes/resigns first responder. Driven from these
-    /// overrides rather than `NSTextFieldDelegate`'s controlTextDidBeginEditing/EndEditing —
-    /// see the note at the `onFocusChange` assignment in `ScrubbableColorField.updateNSView`.
-    var onFocusChange: ((Bool) -> Void)?
-
-    private var dragOrigin: Double?
-
-    // The default NSTextFieldCell intrinsic size proved unreliable once `.fixedSize()` queried
-    // it eagerly (fields collapsed to ~0pt wide) — compute it directly from the string and font
-    // instead of trusting the cell's own layout pass.
-    override var intrinsicContentSize: NSSize {
-        let currentFont = font ?? NSFont.systemFont(ofSize: NSFont.systemFontSize)
-        let size = (stringValue as NSString).size(withAttributes: [.font: currentFont])
-        return NSSize(width: ceil(size.width) + 2, height: ceil(size.height))
-    }
-
-    override func becomeFirstResponder() -> Bool {
-        // AppKit's own `_setUpFirstResponder`/`_selectFirstKeyView` auto-focuses the first key
-        // view in the window while it's still being ordered onto screen — before it's key —
-        // which would open an edit session (and select-all) on the hue field before the user
-        // has clicked anything. Reject that call outright; genuine focus (click, Tab, or our
-        // own `updateNSView` reconciliation) only ever happens once the window is already key.
-        guard window?.isKeyWindow == true else { return false }
-        let result = super.becomeFirstResponder()
-        if result {
-            if let editor = currentEditor() {
-                let length = (editor.string as NSString).length
-                editor.selectedRange = NSRange(location: 0, length: length)
-                // The shared field editor's default `textContainerInset` doesn't match the metrics
-                // the static `NSTextFieldCell` used to draw the same string, so the text visibly
-                // jumps by a device pixel the instant editing begins. Zero it so the editor draws
-                // the text in exactly the same place the cell did.
-                if let textView = editor as? NSTextView {
-                    textView.textContainerInset = .zero
-                }
-            }
-            window?.invalidateCursorRects(for: self)
-            onFocusChange?(true)
-        }
-        return result
-    }
-
-    override func resignFirstResponder() -> Bool {
-        let result = super.resignFirstResponder()
-        if result {
-            window?.invalidateCursorRects(for: self)
-            onFocusChange?(false)
-        }
-        return result
-    }
-
-    override func viewDidMoveToWindow() {
-        super.viewDidMoveToWindow()
-        window?.invalidateCursorRects(for: self)
-        // Without this, AppKit's own `_setUpFirstResponder`/`_selectFirstKeyView` auto-focuses
-        // the first key view in the window (i.e. this field, if it's first in the hierarchy) the
-        // moment the window becomes key — opening an edit session on the hue field before the
-        // user has clicked anything. `becomeFirstResponder` reports that focus like any other
-        // (correctly, so genuine focus changes stay in sync), so SwiftUI accepts it and shows the
-        // outline. Steer AppKit's auto-pick to the content view instead, which never becomes an
-        // editing session. Harmless to set repeatedly (once per field that attaches).
-        if let window, window.initialFirstResponder !== window.contentView {
-            window.initialFirstResponder = window.contentView
-        }
-    }
-
-    // NSTextField's own `resetCursorRects()` covers `bounds` with an I-beam cursor rect, which
-    // wins over the SwiftUI `.onHover`-driven `NSCursor.set()` the moment the mouse enters this
-    // AppKit view — that's why the resize cursor from the parent's hover handling was reverting
-    // to `|` over the text itself. Claim the rect ourselves while draggable and not being edited.
-    override func resetCursorRects() {
-        guard isDraggable, currentEditor() == nil else {
-            super.resetCursorRects()
-            return
-        }
-        addCursorRect(bounds, cursor: .resizeLeftRight)
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        guard isDraggable else {
-            super.mouseDown(with: event)
-            return
-        }
-
-        let startPoint = event.locationInWindow
-        var didBeginDrag = false
-        let threshold: CGFloat = 2
-
-        // Every exit path below must resolve exactly one of "focus" or a paired
-        // onDragBegin/onDragEnd — an orphaned "began but never ended" session would leave the
-        // parent's edit session stuck open, making the next interaction silently join it instead
-        // of starting fresh.
-        while true {
-            guard let next = NSApp.nextEvent(
-                matching: [.leftMouseDragged, .leftMouseUp],
-                until: .distantFuture,
-                inMode: .eventTracking,
-                dequeue: true
-            ) else {
-                if didBeginDrag { finishDrag() }
-                return
-            }
-
-            switch next.type {
-            case .leftMouseDragged:
-                let translationX = next.locationInWindow.x - startPoint.x
-                if !didBeginDrag {
-                    guard abs(translationX) >= threshold else { continue }
-                    didBeginDrag = true
-                    beginDrag()
-                }
-                updateDrag(translationX: translationX)
-            case .leftMouseUp:
-                if didBeginDrag {
-                    finishDrag()
-                } else {
-                    // A genuine click: focus normally, exactly like a plain click on any
-                    // ordinary text field would. AppKit's own event routing
-                    // (`_handleMouseDownEvent:` → `NSTextFieldCell _selectOrEdit:`) already
-                    // focuses the field before this override even runs — calling
-                    // `makeFirstResponder` again here forces a redundant resign/become pair
-                    // that corrupts the field editor's begin-editing bookkeeping, so
-                    // `controlTextDidBeginEditing` silently never fires and the field never
-                    // reports itself focused to SwiftUI (no outline, no edit session). Once
-                    // focused, first responder is the *field editor* (an NSTextView), not this
-                    // control itself, so check against `currentEditor()` rather than `self`.
-                    let editorIsActive = currentEditor() != nil && window?.firstResponder === currentEditor()
-                    if !editorIsActive {
-                        window?.makeFirstResponder(self)
-                    }
-                }
-                return
-            default:
-                if didBeginDrag { finishDrag() }
-                return
-            }
-        }
-    }
-
-    private func beginDrag() {
-        dragOrigin = Double(stringValue.trimmingCharacters(in: .whitespaces)) ?? 0
-        NSCursor.resizeLeftRight.set()
-        // AppKit's own event routing (`_handleMouseDownEvent:` → `NSTextFieldCell
-        // _selectOrEdit:`) already focused and select-all'd this field as part of routing the
-        // mouseDown that's turning out to be this drag — before this override's loop could
-        // tell click from drag apart. Collapse that selection now that we know it's a drag, so
-        // releasing the mouse doesn't leave the dragged-to value shown text-selected. Just the
-        // selection, not the focus itself — resigning first responder here would race the
-        // deferred focus-loss handling in `EditableColorValue` and could end the drag's edit
-        // session (commit/revert) while the user is still mid-drag.
-        if let editor = currentEditor() {
-            editor.selectedRange = NSRange(location: (editor.string as NSString).length, length: 0)
-        }
-        onDragBegin?()
-    }
-
-    private func updateDrag(translationX: CGFloat) {
-        guard let origin = dragOrigin else { return }
-        let fine = NSEvent.modifierFlags.contains(.option)
-        let unitsPerStep = dragUnitsPerPixel(for: range)
-        var newValue = origin + Double(translationX) * unitsPerStep * (fine ? 0.1 : 1.0)
-        if let range {
-            newValue = min(max(newValue, range.lowerBound), range.upperBound)
-        }
-        onDragChanged?(newValue)
-    }
-
-    private func finishDrag() {
-        dragOrigin = nil
-        NSCursor.arrow.set()
-        onDragEnd?()
     }
 }
